Trait capnp_rpc::OutgoingMessage

pub trait OutgoingMessage {
    fn get_body<'a>(&'a mut self) -> Result<Builder<'a>>;

    fn get_body_as_reader<'a>(&'a self) -> Result<Reader<'a>>;

    fn send(
        self: Box<Self>
    ) -> (Promise<Rc<Builder<HeapAllocator>>, Error>, Rc<Builder<HeapAllocator>>);

    fn take(self: Box<Self>) -> Builder<HeapAllocator>;
}

Required Methods

fn get_body<'a>(&'a mut self) -> Result<Builder<'a>>

fn get_body_as_reader<'a>(&'a self) -> Result<Reader<'a>>

fn send(
self: Box<Self>
) -> (Promise<Rc<Builder<HeapAllocator>>, Error>, Rc<Builder<HeapAllocator>>)

Sends the message. Returns a promise for the message that resolves once the send has completed. Dropping the returned promise does not cancel the send.

fn take(self: Box<Self>) -> Builder<HeapAllocator>